requirements.txt
flask==3.0.2
flask_cors == 4.0.0
mysql-connector_python == 8.3.0
